More backend work

+++ backend/team-registration/src/main/java/app/repository/PlayerRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import app.model.Player;
 
-public interface PlayerRepository extends CrudRepository<Player, Integer> {
+public interface PlayerRepository extends CrudRepository<Player, Long> {
 
 }
